refactor: move Message related classes
- Move Message in events as it is primarily an event.
- rename MessageDispatcher into Dispatcher and move it into nex dispatch
  package along MessageHandler and DispatchException
