+++ src/main/java/org/sar/ppi/peersim/PeerSimInfrastructure.java
@@ -4,8 +4,8 @@ import java.lang.reflect.InvocationTargetException;
 import org.sar.ppi.Infrastructure;
 import org.sar.ppi.NodeProcess;
 import org.sar.ppi.Ppi;
-import org.sar.ppi.communication.Message;
 import org.sar.ppi.events.Event;
+import org.sar.ppi.events.Message;
 import peersim.config.Configuration;
 import peersim.core.Network;
 import peersim.core.Node;
